Start interpreter first, then spawn LSP servers.
So far our strategy was as follows:

- upon loading a scheme file, start a command server in background.
- new files communicate to the command server, requesting for LSP servers on
  specific ports.
- (lsp-run-scheme) connects to a REPL server running in the command server.

The problem with this approach is that the user is expected to interact with
this remote REPL directly. This works fine with guile, for CHICKEN though,
the REPL experience is limited, since for example typical csi commands
are missing.

The new strategy is now as follows:

- the user is expected to start the interpreter first, by running lsp-scheme-run. This starts either csi or guile. This also starts a command server in the running REPL, which listens on a port provided by the user.
- upon visiting a scheme file, a request is sent through TCP to the command server, spawning LSP server on specific ports.

Unfortunately I still didn't find a way to allow launching the REPL when opening a scheme file. This should be fixed next.