Join files into lsp-scheme.el; add missing script
Since packaging rules demand that each package has its name as prefix, and
there were circular dependencieas between lsp-guile/lsp-chicken and
lsp-scheme, I got rid of the split and put everything in lsp-scheme.

The alternative would be to define recipes for lsp-chicken and lsp-guile. I'm
not sure that this would be productive, especially because there is not much
code needed for specific implementations. Also I plan to add support to more
implementations in the future.