Delay client registration
If more then one lsp-client is registered, the user cannot define through
customization which one will be used. So you delay client registration until
lsp-scheme is called.

Important change: #'lsp hook must not be used, use #'lsp-scheme instead. It
will automatically #'lsp when started. This change is due to the registration
delay. If #'lsp is added as hook to scheme-mode, it will complain that no
clients are available.