Major design changes
We don't couple the LSP server to the REPL anymore. Instead we simply call
the executable lsp-server (chicken-lsp-server or guile-lsp-server), as
other LSP-based clients do.

This change simplies the extension. The disadvantage for now is that we lose
"jump to definition" in Guile, since the user can't load a file in the REPL
and update the runtime. I will investigate an workaround for that, for example
by allowing loading a file through a custom LSP command (as vscode-scheme-lsp
does).