remove auto-install of chicken; refactor gambit-lsp-server install
We don't auto-install the LSP server for CHICKEN anymore. Installing it
locally to the extension is suboptimal, since many dependencies can be fetched
and installed. Also installing chicken-lsp-server is trivial, so the user
can install it manually.

Installation of the LSP server for gambit now has two separate checks. It installs
library + executable if the first is not found. If the library is found but
only the executable is missing, it only compiles the executable and reuses
the installed library.